pyproject.toml
[tool.poetry]
name = "legends of runeterra deckroll"
version = "1.0.0"
description = ""
authors = ["Ghostdriver <mattes.hupp@t-online.de>"]

[tool.poetry.dependencies]
python = "^3.10"
lor-deckcodes = "^5.0.0"
XlsxWriter = "^3.0.3"
requests = "^2.28.2"
"discord.py" = "^2.1.1"

[tool.poetry.dev-dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
